+++ tailwind.config.js
@@ -4,7 +4,11 @@ module.exports = {
   ],
   darkMode: false, // or 'media' or 'class'
   theme: {
-    extend: {},
+    extend: {
+      lineHeight: {
+        '0': '0',
+       }
+    },
   },
   variants: {
     extend: {},